[Refactor][CodeGen] Refactor CodeGen part for multi-backend decoupling
