+++ src/backend/cuda/codegen/rt_mod_cuda.cc
@@ -1,8 +1,8 @@
-#include "../transform/common/attr.h"
 #include "codegen_cuda.h"
 #include "runtime/cuda/cuda_module.h"
 #include "runtime/meta_data.h"
 #include "runtime/pack_args.h"
+#include "transform/common/attr.h"
 #include <tvm/ffi/reflection/registry.h>
 #include <tvm/ir/transform.h>
