+++ src/backend/metal/CMakeLists.txt
@@ -11,7 +11,7 @@ if(NOT APPLE)
 endif()
 
 file(GLOB TILE_LANG_METAL_SRCS
-  src/target/rt_mod_metal.cc
+  src/backend/metal/codegen/rt_mod_metal.cc
 )
 list(APPEND TILE_LANG_SRCS ${TILE_LANG_METAL_SRCS})
 # FIXME: CIBW failed with backtrace, why???
