+++ src/backend/rocm/codegen/intrin_rule_hip.cc
@@ -5,7 +5,7 @@
 #include <tvm/tir/builtin.h>
 #include <tvm/tir/op_attr_types.h>
 
-#include "../support/ffi_aliases.h"
+#include "support/ffi_aliases.h"
 #include "target/intrin_rule.h"
 
 namespace tvm {
