+++ src/backend/rocm/codegen/rt_mod_hip.cc
