+++ src/backend/rocm/codegen/stubs/hip.cc
