+++ src/backend/rocm/codegen/stubs/hip.h
