+++ src/backend/rocm/codegen/stubs/hiprtc.cc
