+++ src/op/utils.h
@@ -6,8 +6,8 @@
 #ifndef TVM_TL_OP_UTILS_H_
 #define TVM_TL_OP_UTILS_H_
 
-#include "../target/stubs/cuda.h"
 #include "./operator.h"
+#include "backend/cuda/codegen/stubs/cuda.h"
 #include "region.h"
 #include "tvm/runtime/base.h"
 #include <tvm/tir/buffer.h>
